[#367] Refactor signup code to be reusable
