+++ frontend/src/modules/signup/form.jsx
@@ -1,4 +1,3 @@
-import { signUpData, initialSignUpData } from "../signup/view";
 import React, { useEffect, useState, useCallback } from "react";
 import { notification } from "antd";
 import { withTheme } from "@rjsf/core";
@@ -20,8 +19,10 @@ import {
   checkDependencyAnswer,
   customFormats,
 } from "../../utils/forms";
-
-import uiSchema from "./signUiSchema.json";
+import entity from "./entity";
+import stakeholder from "./stakeholder";
+import entityUiSchema from "./entityUiSchema.json";
+import stakeholderUiSchema from "./stakeholderUiSchema.json";
 
 import { UIStore } from "../../store";
 import { withRouter } from "react-router-dom";
@@ -34,6 +35,7 @@ const SignUpForm = withRouter(
     sending,
     setSending,
     representEntity,
+    formType,
     highlight,
     setHighlight,
     formSchema,
@@ -43,6 +45,10 @@ const SignUpForm = withRouter(
   }) => {
     const { countries, organisations, tags, formEdit } = UIStore.currentState;
     const { status, id } = formEdit.signUp;
+    const { initialSignUpData, signUpData } =
+      formType === "entity" ? entity : stakeholder;
+    const uiSchema =
+      formType === "entity" ? entityUiSchema : stakeholderUiSchema;
     const signUpFormData = signUpData.useState();
     const [dependValue, setDependValue] = useState([]);
     const [editCheck, setEditCheck] = useState(true);
