+++ frontend/src/modules/signup/stakeholderUiSchema.json
@@ -1,0 +1,268 @@
+{
+  "S1": {
+    "classNames": "list-group",
+    "titleAndLastName": {
+      "ui:options": {
+        "orderable": false
+      },
+      "title": {
+        "ui:widget": "select",
+        "ui:span": 5,
+        "ui:placeholder": "Title"
+      },
+      "lastName": {
+        "ui:span": 19,
+        "ui:placeholder": "Last name"
+      }
+    },
+    "firstName": {
+      "ui:placeholder": "First name"
+    },
+    "email": {
+      "ui:widget": "email",
+      "ui:placeholder": "Email",
+      "ui:disabled": true
+    },
+    "publicEmail": {
+      "ui:widget": "checkbox",
+      "ui:label": null
+    },
+    "linkedin": {
+      "ui:placeholder": "Your LinkedIn username",
+      "ui:widget": "TextIconWidget",
+      "ui:icon": "linkedin"
+    },
+    "twitter": {
+      "ui:placeholder": "Your twitter username",
+      "ui:widget": "TextIconWidget",
+      "ui:icon": "twitter"
+    },
+    "photo": {
+      "ui:options": { "accept": "image/*" },
+      "ui:widget": "file"
+    },
+    "representativeSector": {
+      "ui:widget": "select",
+      "ui:placeholder": "Search sectors",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+    "country": {
+      "ui:widget": "select",
+      "ui:placeholder": "Search country",
+      "ui:showSearch": true
+    },
+    "geoCoverageType": {
+      "ui:placeholder": "Choose the coverage type",
+      "ui:widget": "select"
+    },
+    "geoCoverageValueRegional": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+    "geoCoverageValueNational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true
+    },
+    "geoCoverageValueTransnational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+    "geoCoverageValueSubNational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true
+    },
+    "geoCoverageValueGlobalSpesific": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+
+    "S1_ExpertisesAndActivities": {
+      "classNames": "list-group-item",
+      "seeking": {
+        "ui:widget": "select",
+        "ui:placeholder": "Search",
+        "ui:mode": "multiple",
+        "ui:showSearch": true
+      },
+      "offering": {
+        "ui:widget": "select",
+        "ui:placeholder": "Search",
+        "ui:mode": "multiple",
+        "ui:showSearch": true
+      },
+      "about": {
+        "ui:placeholder": "",
+        "ui:widget": "textarea"
+      },
+      "tags": {
+        "ui:widget": "select",
+        "ui:placeholder": "Search tags",
+        "ui:showSearch": true,
+        "ui:mode": "multiple"
+      },
+      "portfolio": {
+        "ui:widget": "file",
+        "ui:options": {
+          "accept": ".doc,.docx,.xml,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document,.pdf,application/pdf"
+        }
+      },
+      "public_database": {
+        "ui:widget": "checkbox",
+        "ui:label": null
+      },
+      "ui:order": [
+        "seeking",
+        "offering",
+        "about",
+        "tags",
+        "portfolio",
+        "public_database"
+      ]
+    },
+    "ui:order": [
+      "titleAndLastName",
+      "firstName",
+      "email",
+      "publicEmail",
+      "linkedin",
+      "twitter",
+      "photo",
+      "representativeSector",
+      "country",
+      "geoCoverageType",
+      "geoCoverageValueRegional",
+      "geoCoverageValueNational",
+      "geoCoverageValueTransnational",
+      "geoCoverageValueSubNational",
+      "geoCoverageValueGlobalSpesific",
+      "S1_ExpertisesAndActivities"
+    ]
+  },
+  "S2": {
+    "classNames": "list-group",
+    "orgName": {
+      "ui:placeholder": "Organisation name"
+    },
+    "orgRepresentative": {
+      "ui:widget": "radio"
+    },
+    "orgDescription": {
+      "ui:placeholder": "Organisartion brief",
+      "ui:widget": "textarea"
+    },
+    "orgUrl": {
+      "ui:addOnBefore": "https://",
+      "ui:widget": "URLWidget",
+      "ui:placeholder": "Url"
+    },
+    "orgLogo": {
+      "ui:options": { "accept": "image/*" },
+      "ui:widget": "file"
+    },
+    "ui:order": [
+      "orgName",
+      "orgRepresentative",
+      "orgDescription",
+      "orgUrl",
+      "orgLogo"
+    ]
+  },
+  "S3": {
+    "classNames": "list-group",
+    "orgExpertise": {
+      "ui:placeholder": "Search",
+      "ui:widget": "select",
+      "ui:mode": "multiple",
+      "ui:showSearch": true
+    },
+    "orgContribution": {
+      "ui:widget": "checkboxes",
+      "ui:label": null
+    },
+    "ui:order": ["orgExpertise", "orgContribution"]
+  },
+  "S4": {
+    "ui:order": [
+      "orgHeadquarter",
+      "orggeoCoverageType",
+      "orggeoCoverageValueRegional",
+      "orggeoCoverageValueNational",
+      "orggeoCoverageValueTransnational",
+      "orggeoCoverageValueSubNational",
+      "orggeoCoverageValueGlobalSpesific"
+    ],
+    "orgHeadquarter": {
+      "ui:widget": "select",
+      "ui:placeholder": "Search",
+      "ui:showSearch": true
+    },
+    "orggeoCoverageType": {
+      "ui:placeholder": "Choose the coverage type",
+      "ui:widget": "select"
+    },
+    "orggeoCoverageValueRegional": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+    "orggeoCoverageValueNational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true
+    },
+    "orggeoCoverageValueTransnational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    },
+    "orggeoCoverageValueSubNational": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true
+    },
+    "orggeoCoverageValueGlobalSpesific": {
+      "ui:placeholder": "Choose the coverage",
+      "ui:widget": "select",
+      "ui:showSearch": true,
+      "ui:mode": "multiple"
+    }
+  },
+  "S5": {
+    "ui:order": ["registeredStakeholders", "otherStakeholders"],
+    "registeredStakeholders": {
+      "ui:widget": "select",
+      "ui:placeholder": "Search by name or email",
+      "ui:mode": "multiple",
+      "uniqueItems": true,
+      "ui:showSearch": true
+    },
+    "otherStakeholders": {
+      "ui:options": {
+        "addItemText": "Add another stakeholder",
+        "removeItemText": "Remove stakeholder",
+        "orderable": false
+      },
+      "classNames": "list-group-item",
+      "ui:group": "border",
+      "items": {},
+      "ui:order": [
+        "title",
+        "lastName",
+        "firstName",
+        "email",
+        "organisationRole"
+      ]
+    }
+  }
+}
